run-qemu: make config read failure fatal
There's no way to encapsulate the check into a function. Won't
work: set -e, rewriting in terms of compare function with return code,
explicit exit from the function. Because of two reasons:

1. $() invokes a nested shell, so exiting from the function won't
kill the parent shell

2. Neither "$()=value" or "if syscfg_cmp field value" will fail when
the inner command fails, even with set -e. The latter for obvious
reason, and the former $() returns the exit code, but the way
it's used (string comparison or assigning to a variable) won't
fail the parent script despite the non-zero error code and set -e.