zebu: memory map
Kernel image (uImage) should be generated by the build system, not by
run-qemu.sh at time of run. The top-level makefile in hpsc parent
repo does that, the path here is where that makefile puts the image.

If need be, we can introduce environment variables into the memory map
definition. We already have qemu-env.sh for the *run* environment, so we
could also add a build-env.sh for the *build* environment.  Until we
need to, though, keep things simple.