ssw: conf: base: busybox: math,tar,stat,du,df
Needed by various "tool" profiles. If we care about size, we
could add this to those profiles only, instead of adding to the base,
but the trade-off is that then we would need to recompile busybox
every time we run a different profile (every time we run the "tool"),
which might be ok -- incremental recompile will be fast.