prof: lib-trch-bm: decrease size of boot SFS
LSIO SMC SRAM max size in Chiplet Mem Map v14 is 16Mb per chip, due to the
address region allocated to "LSIO SMC SRAM/NAND" = 128Mb (64M for NAND and 16x4
MB for SRAM is the maximum power of two mem size combination).

And, we can't span the file system image across chips, because the first
section of each chip is used to store BL0 blob and BL1 image. In Mem Map v16,
SRAM is not