prof: hpps-linux: serial to 500000 and set in init
Need to set the baudrate from init because at the end of the boot,
when the kernel opens /dev/console, it (unconditionally) applies
the default 9600 baudrate set in serial_core.c in init_termios.

Switching to 500000 because busybox stty has a table of baudrates
and 125000 is not one of them, so stty fails. Luckily, 500000 is
in the table and is a multiple of 16MHz clock used in Zebu.

Add stty to busybox config.